Задание 3. Автотесты для UI. Исправил почти все замечания.

+++ src/test/java/ru/practicum/AuthorizationTest.java
@@ -1,5 +1,6 @@
 package ru.practicum;
 
+import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
 import org.apache.commons.lang3.RandomStringUtils;
@@ -7,12 +8,10 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import ru.practicum.pageObject.*;
+import ru.practicum.page.object.*;
 import ru.practicum.model.User;
 import ru.practicum.steps.ClientUser;
 import ru.practicum.utils.Urls;
-
-
 
 public class AuthorizationTest extends BaseTest {
     HomePage homePage;
@@ -20,7 +19,6 @@ public class AuthorizationTest extends BaseTest {
     ProfilePage profilePage;
     RegistrationPage registrationPage;
     RestorePasswordPage restorePasswordPage;
-
 
     String email;
     String password;
@@ -30,7 +28,7 @@ public class AuthorizationTest extends BaseTest {
     User user;
 
     @Before
-    public void createUser(){
+    public void createUser() {
         name = RandomStringUtils.randomAlphabetic(7);
         email = RandomStringUtils.randomAlphanumeric(6) + "@yandex.ru";
         password = RandomStringUtils.randomAlphabetic(7);
@@ -41,6 +39,7 @@ public class AuthorizationTest extends BaseTest {
 
     @Test
     @DisplayName("Проверка входа в аккаунт через нажатие кнопки 'Войти' на домашней странице")
+    @Description("Авторизация через главную страницу: ввод данных и проверка попадания в личный профиль")
     public void enterAccountByClickingEnterAccountButtonOnHomePageTest() {
         homePage = new HomePage(webDriver);
         homePage.waitForEnterAccountButton();
@@ -57,8 +56,10 @@ public class AuthorizationTest extends BaseTest {
         Assert.assertEquals(email.toLowerCase(), profilePage.getEmailText().toLowerCase());
     }
 
+
     @Test
     @DisplayName("Вход через кнопку «Личный кабинет»")
+    @Description("Переход к авторизации через ссылку в личном кабинете и последующее попадание в профиль")
     public void loginViaPersonalProfileButtonTest(){
         homePage = new HomePage(webDriver);
         homePage.waitForPersonalProfileButton();
@@ -76,6 +77,7 @@ public class AuthorizationTest extends BaseTest {
 
     @Test
     @DisplayName("Вход через кнопку в форме регистрации")
+    @Description("Авторизация через форму регистрации: быстрый переход к авторизации и последующий вход в профиль")
     public void loginViaRegistrationFormLinkTest(){
         homePage = new HomePage(webDriver);
         homePage.waitForEnterAccountButton();
@@ -98,6 +100,7 @@ public class AuthorizationTest extends BaseTest {
 
     @Test
     @DisplayName("Вход через кнопку в форме восстановления пароля")
+    @Description("Проверка переход к авторизации из окна восстановления пароля и успешный вход в профиль")
     public void loginUsingRestorePasswordFormLinkTest(){
         homePage = new HomePage(webDriver);
         homePage.waitForEnterAccountButton();
@@ -119,7 +122,9 @@ public class AuthorizationTest extends BaseTest {
     }
 
     @After
-    public void cleanUp(){
-        userClient.deleteUser(accessToken);
+    public void cleanUp() {
+        if (accessToken != null && !accessToken.isEmpty()) {
+            userClient.deleteUser(accessToken);
+        }
     }
 }
